@levanhoangphuong

+++ raspberry/main/main.c
@@ -21,13 +21,13 @@
 int main(int argc, char const *argv[])
 {
 	/*platform thread: control os_shell_thread & embedded_shell_thread*/
-	void *thread_exit; //message when thread cancelled
-	pthread_t platform_thread; //declare thread name
+	void *thread_exit; 		//message when thread cancelled
+	pthread_t platform_thread; 	//declare thread name
 
-	createThread(&platform_thread, platform, thread_exit);	
+	if(!createThread(&platform_thread, platform, thread_exit)) exit(1);
 	pthread_join(platform_thread, &thread_exit);
-	fprintf(stderr, "%s\n", (char *) thread_exit);
+	fprintf(stderr, "%s\n", (char *)thread_exit);
 
-	unlink("main"); //remove main.o
+	unlink("main"); 			//remove main.o
 	return 1;
 }
